test(e2e): Move support elements to support folder
