+++ test/e2e/main/cli-arguments.spec.js
@@ -9,7 +9,8 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const path = require("path");
-const { CliRunner, request, wait, TimeCounter, BINARY_PATH } = require("./utils");
+const { request, wait, TimeCounter, BINARY_PATH } = require("../support/utils");
+const CliRunner = require("../support/CliRunner");
 
 describe("command line arguments", () => {
   const cwdPath = path.resolve(__dirname, "fixtures");
