+++ test/e2e/mocks-server
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 "use strict";
 
-require("./start").start();
+require("./support/start").start();
 
 /*
 Copyright 2019 XbyOrange
