+++ test/e2e/support/InteractiveCliRunner.js
@@ -8,7 +8,8 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const { CliRunner, wait } = require("./utils");
+const { wait } = require("./utils");
+const CliRunner = require("./CliRunner");
 
 const LOG = "[CLI]: ";
 const SCREEN_SEPARATOR = ">> Mocks server";
