+++ test/e2e/support/start.js
@@ -12,7 +12,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 "use strict";
 
 const { Core } = require("@mocks-server/core");
-const InquirerCli = require("../../index");
+const InquirerCli = require("../../../index");
 
 const handleError = (error) => {
   console.error(`Error: ${error.message}`);
